// src/app/portal/abilities/models/AbilityActionType.model.ts
export enum AbilityActionType {
  CLASSROOM = 'CLASSROOM',
  ASCENSION = 'ASCENSION',
  THEFT = 'THEFT',
  REVEAL = 'REVEAL',
  HEALTH = 'HEALTH',
  REVIVE = 'REVIVE',
  DEFEREAL_HOMEWORK = 'DEFEREAL_HOMEWORK',
  PROTECTION = 'PROTECTION',
  MIRROR = 'MIRROR',
  REVENGE = 'REVENGE',
  INVULNERABILITY = 'INVULNERABILITY'
}